// types/components/card.ts
export type cardIcons =
  | "user"
  | "brain"
  | "heart"
  | "search"
  | "file"
  | "chat"
  | "star"
  | "book"
  | "lightbulb"
  | "balance-scale"
  | "chart-bar"
  | "frown"
  | "check"
  | "shopping-cart"
  | "exclamation-triangle"
  | "life-ring"
  | "clock"
  | "cogs"
  | "users"
  | "comments"
  | "redo"
  | "refresh"
  | "gift"
  | "ban"
  | "heart"
  | "check-square"
  | "frown"
  | "star"
  | "chat"
  | "book"
  | "search"
  | "balance-scale"
  | "chart-bar"
  | "star"
  | "book"
  | "chat"
  | "frown"
  | "check"
  | "life-ring"
  | "clock"
  | "redo"
  | "refresh"
  | "life-ring"
  | "gift"
  | "frown"
  | "comments"
  | "heart";